Move SetFlag() with bool functionality to new ToggleFlag().
Now SetFlag() and UnsetFlag() always only ever set or unset a
flag. There is a new ToggleFlag() method with two overloads. With a
boolean, it will call SetFlag() or UnsetFlag() appropriately for
you. Without a boolean it will unset the flag if it is set or set it
if it is unset. According to @arogozine, this is less confusing than
being able to set a flag to OFF with SetFlag().